Change cluster_id parameters from uint8_t to uint16_t
This is needed to allow the window covering cluster to be used.

+++ src/silabs/hal/zigbee.c
@@ -117,7 +117,7 @@ void hal_zigbee_init(hal_zigbee_endpoint *endpoints, uint8_t endpoints_cnt) {
   }
 }
 
-void hal_zigbee_notify_attribute_changed(uint8_t endpoint, uint8_t cluster_id,
+void hal_zigbee_notify_attribute_changed(uint8_t endpoint, uint16_t cluster_id,
                                          uint16_t attribute_id) {
   hal_zigbee_cluster *cluster = find_hal_cluster(endpoint, cluster_id);
   hal_zigbee_attribute *attr =
